+++ RedPandaIDE/HighlighterManager.cpp
@@ -19,6 +19,7 @@
 #include <QObject>
 #include "qsynedit/highlighter/cpp.h"
 #include "qsynedit/highlighter/asm.h"
+#include "qsynedit/highlighter/glsl.h"
 #include "qsynedit/Constants.h"
 #include "colorscheme.h"
 
@@ -42,6 +43,8 @@ PSynHighlighter HighlighterManager::getHighlighter(const QString &filename)
                 || suffix == "CPP" || suffix =="H" || suffix == "c++"
                 || suffix == "h++") {
             return getCppHighlighter();
+        } else if (suffix == "vs" || suffix == "fs") {
+            return getGLSLHighlighter();
         }
     }
     return PSynHighlighter();
@@ -97,6 +100,34 @@ PSynHighlighter HighlighterManager::getAsmHighlighter()
     highlighter->whitespaceAttribute()->setForeground(Qt::lightGray);
     highlighter->stringAttribute()->setForeground(0x007d17);
     highlighter->symbolAttribute()->setForeground(0xc10000);
+    return pHighlighter;
+}
+
+PSynHighlighter HighlighterManager::getGLSLHighlighter()
+{
+    SynEditGLSLHighlighter* highlighter = new SynEditGLSLHighlighter();
+    PSynHighlighter pHighlighter(highlighter);
+    highlighter->asmAttribute()->setForeground(Qt::blue);
+    highlighter->charAttribute()->setForeground(Qt::black);
+    highlighter->commentAttribute()->setForeground(0x8C8C8C);
+    highlighter->commentAttribute()->setStyles(SynFontStyle::fsItalic);
+    highlighter->classAttribute()->setForeground(0x008080);
+    highlighter->floatAttribute()->setForeground(Qt::darkMagenta);
+    highlighter->functionAttribute()->setForeground(0x00627A);
+    highlighter->globalVarAttribute()->setForeground(0x660E7A);
+    highlighter->hexAttribute()->setForeground(Qt::darkMagenta);
+    highlighter->identifierAttribute()->setForeground(0x080808);
+    highlighter->invalidAttribute()->setForeground(Qt::red);
+    highlighter->localVarAttribute()->setForeground(Qt::black);
+    highlighter->numberAttribute()->setForeground(0x1750EB);
+    highlighter->octAttribute()->setForeground(Qt::darkMagenta);
+    highlighter->preprocessorAttribute()->setForeground(0x1f542e);
+    highlighter->keywordAttribute()->setForeground(0x0033b3);
+    highlighter->whitespaceAttribute()->setForeground(Qt::lightGray);
+    highlighter->stringAttribute()->setForeground(0x007d17);
+    highlighter->stringEscapeSequenceAttribute()->setForeground(Qt::red);
+    highlighter->symbolAttribute()->setForeground(0xc10000);
+    highlighter->variableAttribute()->setForeground(0x400080);
     return pHighlighter;
 }