+++ RedPandaIDE/HighlighterManager.h
@@ -27,6 +27,7 @@ public:
     PSynHighlighter copyHighlighter(PSynHighlighter highlighter);
     PSynHighlighter getCppHighlighter();
     PSynHighlighter getAsmHighlighter();
+    PSynHighlighter getGLSLHighlighter();
     void applyColorScheme(PSynHighlighter highlighter, const QString& schemeName);
 };